make return value of scoreing and validator a dict
To be able to return additional scoreing and validation
information a dict is much more readable than a tuple.

The dict returned by a validator must contain a 'status'
key, the dict returned by a scoreing must contain a
'score' key. All other keys are specific to the strategy
used.