catch Errors on flushInput
After opening the serial port the SIReader trys to flush available
input before checking if an SI Station is connected to the serial port.
This fails if the serial port is not ready for whatever reason (eg.
there is no real device behind the device node). Catch the resulting
exception and progagate the error condition as an SIReaderException upwards.