refactor run_editor into separate panel classes
The different panels inside the run_editor notebook are now separate
classes. This allows the EditRunPanel and the ReadoutRunPanel to share a
significant amount of code in a common base class.

The EditRunPanel is restructured and editing fields for manual start and
finish time specification are added. This adopts the run_editor to the
changes in the database layout.