Edit function for runner properties
Runner properties are now editable in the EditRunPanel. To allow easier
editing given_name and surname now default to ''. This makes it possible
to create "empty" Runners. Only the basic properties name, dateofbirth,
number and team are editable due to space restrictions in the GUI. Later
additional dialogs to enter full runner details and to add new teams
coudl be added.