remove option to connect courses to start and finish
The special SI-Stations for start and finish are no longer used. Remove
the command line options to connect courses to these stations from
"import". The code to add these stations is still present in the
importer class. It should be remove with a general cleanup of all code
for these special stations.