Add importer for the SOLV runner database
The importer imports all runners from a CSV file downloaded from the
SOLV website. The supported format is called VELPOZ data file on the
site. If an imported runner is already in the database the data for this
runner is updated. SI-Card numbers are never overwritten.

Additionally this adds a verbose option to the import command.