Add registration mode to run_editor
When on the EditRunPanel you can enable a registration mode. In this
mode SI-Cards are read and new runs for this card are created
automatically. This is useful for on the spot registration of runners.

To avoid bad interactions with the normal readout mode this also removes
the "autorun" feature from RunEditor. If you want to load a run from an
SI-card you now have to always call load_run_from_card. The RunEditor
will notify you of new cards inserted by a "reader" event.