Printout of single runs
This adds the possibility to print out the traditional "run recipe":

ReportlabRunFormatter: Creates the actual PDF to print. This uses the
reportlab library.

The run_editor is changed to add a print button on the EditRunPanel, a
checkbox to activate autoprint on the ReadoutPanel and a menu item to
configure the print command.