Use clock_timestamp instead of NOW() for log
The SQL NOW() function returns the same time for every invocation during
a transaction. This leads to entries in the log table not having the
correct time of the change and not being picked up by the
TriggerEventObserver, because they appear to be older than they actually
are. Using clock_timestamp fixes this by returning the actual time at
function execution.

Unfortunately this is a PostgreSQL extension and not part of the SQL
standard.