Change meaning of DID_NOT_START and NOT_COMPLETED
DID_NOT_START means that a runner was on the entry list but did not
compete in the race. NOT_COMPLETED means that a runner is in the entry
list, but no final result is available yet. NOT_COMPLETED does not say
anything about the fact that the runner started or not.