"New style" Validation and Scoreing - Step 1
This moves validation and scoreing configuration closer to the Run,
Runner, Team, Course and Category classes. The idea is that these
classes store their validation and scoreing configuration and in the end
persist this to the database. With this event setup moves from a conf.py
file and a mess of different Event subclasses down to the individual
elements. This should make the Event class much simpler and allow for
more flexible Event setups.

This commit only contains the first step in this direction, allowing
runs to score and validate itself. This is needed to be able to properly
reorder Runs in CombinedCourses.